refactor: Polish home app grouping
