+++ app/src/main/java/app/morphe/manager/domain/bundles/PatchBundleSource.kt
@@ -85,11 +85,21 @@ sealed class PatchBundleSource(
         val PatchBundleSource.isDefault inline get() = uid == 0
         val PatchBundleSource.asRemoteOrNull inline get() = this as? RemotePatchBundle
 
+        /**
+         * Resolved avatar URLs for a source in preference order. [primary] is the URL to try
+         * first; [fallback] is what to load if the primary fails (or null when no second
+         * candidate exists).
+         */
         data class AvatarUrls(
             val primary: String?,
             val fallback: String?
         )
 
+        /**
+         * Pick the best avatar URLs for a source. Preference order is custom bundle avatar,
+         * then GitHub, then GitLab; whichever is chosen as [AvatarUrls.primary], the next
+         * available one becomes [AvatarUrls.fallback].
+         */
         val PatchBundleSource.avatarUrls: AvatarUrls get() {
             val bundle = bundleAvatarUrl
             val github = githubAvatarUrl
